pyproject.toml
[tool.poetry]
name = "lti-provider"
version = "0.1.0"
description = ""
authors = ["Alejandro Latorre Rojas <alej669@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
sqlmodel = "^0.0.22"
python-multipart = "^0.0.19"
python-dotenv = "^1.0.1"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
httpx = { version = "^0.28.1", extras = ["http2"] }
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
cryptography = "^41.0.0"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
